==> 12_2023_Instacart_Basket_Analysis/03_Scripts/Step2_IC_Orders_products_combined.ipynb <==
# Project: 12-2023 Instacart Basket Analysis
## Author: Nadia Ordonez
## Step 2 IC Orders products combined

# Table of Contents
* [1. Introduction](#1.-Introduction)
* [2. Importing data](#2.-Importing-data)
    * [2.1 Importing libraries](#2.1-Importing-libraries)
    * [2.2 Importing data](#2.2-Importing-data)
* [3. Data combining](#3.-Data-combining)
    * [3.1 RAM memory space](#3.1-RAM-memory-space)
    * [3.2 Key variable](#3.2-Key-variable)
    * [3.3 Merge](#3.3-Merge)
* [4. Exporting data](#4.-Exporting-data) 

# 1. Introduction

To answer Instacart research question, all dataframes will be combined. First, the Order dataframe will be combined with the Orders products prior. In this way, each order will be enriched to display all shopping items related to each of the 3421083 total orders. 

# 2. Importing data

## 2.1 Importing libraries

#Import analytical libraries
import pandas as pd
import numpy as np
import os

## 2.2 Importing data

#Project folder path into a string to easily retrieve data
path = r'C:\Users\Ich\Documents\12-2023 Instacart Basket Analysis'

### Order

#Import “orders_step1.csv”
#See "Step 1 IC Data Import, Wrangling and Consistency checks" to check for clean up process
orders = pd.read_csv(os.path.join(path, '02 Data', 'Prepared data', 'orders_step1.csv'), index_col = False)

#Check df size
orders.shape

#Check headers
orders.head()

### Orders products prior

#Import “orders_products_prior_step1.pkl”
#See "Step 1 IC Data Import, Wrangling and Consistency checks" to check for clean up process
orders_products_prior = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared data', 'orders_products_prior_step1.pkl'))

#Check df size
orders_products_prior.shape

#Check headers
orders_products_prior.head()

# 3. Data combining

## 3.1 RAM memory space

### Orders

#RAM memory issues can be avoided converting data types 
#Explore data types in the df
orders.dtypes

#Convert specific columns to more memory-efficient types
orders['order_id'] = orders['order_id'].astype('int32')

#Convert specific columns to more memory-efficient types
orders['user_id'] = orders['user_id'].astype('int32')

#Convert specific columns to more memory-efficient types
orders['order_number'] = orders['order_number'].astype('int8')

#Convert specific columns to more memory-efficient types
orders['orders_day_of_week'] = orders['orders_day_of_week'].astype('int8')

#Convert specific columns to more memory-efficient types
orders['order_hour_of_day'] = orders['order_hour_of_day'].astype('int8')

#See results
orders.dtypes

### Orders products prior

#RAM memory issues can be avoided converting data types 
#Explore data types in the df
orders_products_prior.dtypes

#Convert specific columns to more memory-efficient types
orders_products_prior['order_id'] = orders_products_prior['order_id'].astype('int32')

#Convert specific columns to more memory-efficient types
orders_products_prior['product_id'] = orders_products_prior['product_id'].astype('int32')

#Convert specific columns to more memory-efficient types
orders_products_prior['add_to_cart_sequence'] = orders_products_prior['add_to_cart_sequence'].astype('int32')

#Convert specific columns to more memory-efficient types
orders_products_prior['reordered'] = orders_products_prior['reordered'].astype('int8')

#See results
orders_products_prior.dtypes

## 3.2 Key variable

#The key variable is "order_id". This variable is shared among the two dataframes
#An inner join will join the orders that are present on both dataframes
#Explore "order_id" in Orders
orders['order_id'].describe()
#Orders range from 1 up to 3421083

#Identify the number of unique values in 'order_id' from orders
unique_orders = orders['order_id'].nunique()
print(f"The number of unique values in 'order_id': {unique_orders}")
#Each row corresponds to unique values from 1 to 3421083

#The key variable is "order_id". This variable is shared among the two dataframes
#An inner join will join the orders that are present on both dataframes
#Explore "order_id" in Orders products prior
orders_products_prior['order_id'].describe()
#Orders range from 2 up to 3421083, order_id = 1 is not present.
#Each order_id is not unique at each row within "order_id" variable. 

#Identify the number of unique values in 'order_id' from orders products prior
unique_orders_products_prior = orders_products_prior['order_id'].nunique()
print(f"The number of unique values in 'order_id': {unique_orders_products_prior}")
#It would be expected that there is ONLY information on shopping items for 3214874 orders. 

## 3.3 Merge

#function to merge dataframes
orders_products_combined = orders.merge(orders_products_prior, on = 'order_id')

#See results
orders_products_combined.head()

#Identify the number of unique values in 'order_id' from merged dataframe
unique_orders_products_combined = orders_products_combined['order_id'].nunique()
print(f"The number of unique values in 'order_id': {unique_orders_products_combined}")
#As expected from the previous analyses on the key variable, there is ONLY information on shopping items for 3214874 orders. 

# 4. Exporting data

#check df size
orders_products_combined.shape

#check df headers
orders_products_combined.head()

#Exporting to prepared data folder
#The pickle format is preferred for large df. This df contains 32M rows
orders_products_combined.to_pickle(os.path.join(path, '02 Data','Prepared data', 'orders_products_combined_step2.pkl'))
