==> 12_2023_Instacart_Basket_Analysis/03_Scripts/Step3_IC_Orders_products_merged.ipynb <==
# Project: 12-2023 Instacart Basket Analysis
## Author: Nadia Ordonez
## Step 3 IC Orders products merged

# Table of Contents
* [1. Introduction](#1.-Introduction)
* [2. Importing data](#2.-Importing-data)
    * [2.1 Importing libraries](#2.1-Importing-libraries)
    * [2.2 Importing data](#2.2-Importing-data)
* [3. Data combining](#3.-Data-combining)
    * [3.1 RAM memory space](#3.1-RAM-memory-space)
    * [3.2 Key variable](#3.2-Key-variable)
    * [3.3 Merge](#3.3-Merge)
* [4. Exporting data](#4.-Exporting-data) 

# 1. Introduction

To answer Instacart research question, all dataframes will be combined. Previously, the Order dataframe was combined with the Orders products prior into a single dataframe named as "orders_product_combined". Here, the products dataframe will be additionally combined. In this way, additional product details such as product names and prices will be added.  

# 2. Importing data

## 2.1 Importing libraries

#Import analytical libraries
import pandas as pd
import numpy as np
import os

## 2.2 Importing data

#Project folder path into a string to easily retrieve data
path = r'C:\Users\Ich\Documents\12-2023 Instacart Basket Analysis'

### Order products combined

#Import “orders_products_combined_step2.pkl”
#See "Step 2 IC Orders products combined" to check for merging details
orders_products_combined = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared data', 'orders_products_combined_step2.pkl'))

#Check df size
orders_products_combined.shape

#Check headers
orders_products_combined.head()

### Products

#Import “products_step1.csv”
#See "Step 1 IC Data Import, Wrangling and Consistency checks" to check for clean up process
products = pd.read_csv(os.path.join(path, '02 Data', 'Prepared data', 'products_step1.csv'))

#Check df size
products.shape

#Check headers
products.head()

# 3. Data combining

## 3.1 RAM memory space

### Orders products combined

#RAM memory issues can be avoided converting data types 
#Explore data types in the df
orders_products_combined.dtypes
#Data types were previously converted to save RAM memory (see "Step 2 IC Orders products combined")

### Products

#RAM memory issues can be avoided converting data types 
#Explore data types in the df
products.dtypes

#Convert specific columns to more memory-efficient types
products['product_id'] = products['product_id'].astype('int32')

#Convert specific columns to more memory-efficient types
products['aisle_id'] = products['aisle_id'].astype('int8')

#Convert specific columns to more memory-efficient types
products['department_id'] = products['department_id'].astype('int8')

#See results
products.dtypes

## 3.2 Key variable

#The key variable is "product_id". This variable is shared among the two dataframes
#An inner join will join the orders that are present on both dataframes
#Explore "product_id" in Orders products combined
orders_products_combined['product_id'].describe()
#Products ids range from 1 up to 49688

#Identify the number of unique values in 'product_id' from orders product combined
unique_orders_products_combined = orders_products_combined['product_id'].nunique()
print(f"The number of unique values in 'product_id': {unique_orders_products_combined}")
#If 49688 products id exist, in this df only 49677 shopping items are listed

#The key variable is "product_id". This variable is shared among the two dataframes
#An inner join will join the orders that are present on both dataframes
#Explore "product_id" in products
products['product_id'].describe()
#Product ids range from 1 up to 49688. 

#Identify the number of unique values in 'product_id' from products
unique_products = products['product_id'].nunique()
print(f"The number of unique values in 'product_id': {unique_products}")
#It would be expected that ONLY 49677 unique product ids will be listed after merging both dfs

## 3.3 Merge

#function to merge dataframes
orders_products_merged = orders_products_combined.merge(products, on = 'product_id')

#See results
orders_products_merged.head()

#Identify the number of unique values in 'product_id' from merged dataframe
unique_orders_products_merged = orders_products_merged['product_id'].nunique()
print(f"The number of unique values in 'product_id': {unique_orders_products_merged}")
#As expected from the previous analyses on the key variable, there is ONLY 49677 unique products ids after the merge. 

# 4. Exporting data

#check df size
orders_products_merged.shape

#check df headers
orders_products_merged.head()

#Exporting to prepared data folder
#The pickle format is preferred for large df. This df contains 32M rows
orders_products_merged.to_pickle(os.path.join(path, '02 Data','Prepared data', 'orders_products_merged_step3.pkl'))
