==> 12_2023_Instacart_Basket_Analysis/03_Scripts/Step4_IC_Orders_products_all.ipynb <==
# Project: 12-2023 Instacart Basket Analysis
## Author: Nadia Ordonez
## Step 4 IC Orders products all

# Table of Contents
* [1. Introduction](#1.-Introduction)
* [2. Importing data](#2.-Importing-data)
    * [2.1 Importing libraries](#2.1-Importing-libraries)
    * [2.2 Importing data](#2.2-Importing-data)
* [3. Data combining](#3.-Data-combining)
    * [3.1 RAM memory space](#3.1-RAM-memory-space)
    * [3.2 Key variable](#3.2-Key-variable)
    * [3.3 Merge](#3.3-Merge)
* [4. Exporting data](#4.-Exporting-data) 

# 1. Introduction

To answer Instacart research question, all dataframes will be combined. Previously, the Order dataframe was combined with the Orders products prior into a single dataframe named as "orders_product_combined". Then, the products dataframe was additionally combined and saved as "orders_product_merged". Here, the Customer dataframe will be added. In this way, orders can be linked to customers, and thus customer shopping behaviors can be explored.

# 2. Importing data

## 2.1 Importing libraries

#Import analytical libraries
import pandas as pd
import numpy as np
import os

## 2.2 Importing data

#Project folder path into a string to easily retrieve data
path = r'C:\Users\Ich\Documents\12-2023 Instacart Basket Analysis'

### Order products merged

#Import “orders_products_merged_step3.pkl”
#See "Step 3 IC Orders products merged" to check for merging details
orders_products_merged = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared data', 'orders_products_merged_step3.pkl'))

#Check df size
orders_products_merged.shape

#Check headers
orders_products_merged.head()

### Customers

#Import “customers_step1.csv”
#See "Step 1 IC Data Import, Wrangling and Consistency checks" to check for clean up process
customers = pd.read_csv(os.path.join(path, '02 Data', 'Prepared data', 'customers_step1.csv'))

#Check df size
customers.shape

#Check headers
customers.head()

# 3. Data combining

## 3.1 RAM memory space

### Orders products merged

#RAM memory issues can be avoided converting data types 
#Explore data types in the df
orders_products_merged.dtypes
#Data types were previously converted to save RAM memory (see "Step 3 IC Orders products merged")

### Customers

#RAM memory issues can be avoided converting data types 
#Explore data types in the df
customers.dtypes

#Convert specific columns to more memory-efficient types
customers['user_id'] = customers['user_id'].astype('int32')

#Convert specific columns to more memory-efficient types
customers['gender'] = customers['gender'].astype('category')

#Convert specific columns to more memory-efficient types
customers['state'] = customers['state'].astype('category')

#Convert specific columns to more memory-efficient types
customers['age'] = customers['age'].astype('int32')

#Convert specific columns to more memory-efficient types
customers['number_of_dependants'] = customers['number_of_dependants'].astype('int32')

#Convert specific columns to more memory-efficient types
customers['family_status'] = customers['family_status'].astype('category')

## 3.2 Key variable

#The key variable is "user_id". This variable is shared among the two dataframes
#An inner join will join the orders that are present on both dataframes
#Explore "user_id" in Orders products merged
orders_products_merged['user_id'].describe()
#Users ids range from 1 to 206209 in our df

#Identify the number of unique values in 'user_id' from orders product merged
unique_orders_products_merged = orders_products_merged['user_id'].nunique()
print(f"The number of unique values in 'user_id': {unique_orders_products_merged}")
#There is a total of 206209 unique users ids, starting at 1

#The key variable is "user_id". This variable is shared among the two dataframes
#An inner join will join the orders that are present on both dataframes
#Explore "user_id" in customers
customers['user_id'].describe()
#User ids range from 1 up to 206209 

#Identify the number of unique values in 'user_id' from customers
unique_customers = customers['user_id'].nunique()
print(f"The number of unique values in 'user_id': {unique_customers}")
#It would be expected that we will have customers details for all our users in the orders_products_merged df
#It is also expected to have a similar number of rows are those observed for the orders_products_merged df 

## 3.3 Merge

#function to merge dataframes
orders_products_all = orders_products_merged.merge(customers, on = 'user_id')

#See results
orders_products_all.head()

#Identify the number of unique values in 'user_id' in the resulted dataframe
unique_orders_products_all = orders_products_all['user_id'].nunique()
print(f"The number of unique values in 'user_id': {unique_orders_products_all}")
#As expected from the previous analyses on the key variable, customer information from all users were added. 

#Explore "user_id" in in the resulted dataframe 
orders_products_all['user_id'].describe()
#As expected from the previous analyses on the key variable, 
#the total count of the resulted df equals to the orders_products_merged

# 4. Exporting data

#check df size
orders_products_all.shape

#check df headers
orders_products_all.head()

#Exporting to prepared data folder
#The pickle format is preferred for large df. This df contains 32M rows
orders_products_all.to_pickle(os.path.join(path, '02 Data','Prepared data', 'orders_products_all_step4.pkl'))
