==> 12_2023_Instacart_Basket_Analysis/03_Scripts/Step5_IC_Final_dataset_after_exclusion_flag.ipynb <==
# Project: 12-2023 Instacart Basket Analysis
## Author: Nadia Ordonez
## Step 5 IC Final dataset after exclusion flag

# Table of Contents
* [1. Introduction](#1.-Introduction)
* [2. Importing data](#2.-Importing-data)
    * [2.1 Importing libraries](#2.1-Importing-libraries)
    * [2.2 Importing data](#2.2-Importing-data)
* [3. Exclusion flag](#3.-Exclusion-flag)
    * [3.1 Aggregate variable](#3.1-Aggregate-variable)
    * [3.2 Derived variable](#3.2-Derived-variable)
    * [3.3 Final dataframe](#3.3-Final-dataframe)
* [4. Exporting data](#4.-Exporting-data) 

# 1. Introduction

To answer Instacart research question, all dataframes were combined into a single dataframe. Since the Instacart CFO isn’t interested in customers who don’t generate much revenue for the app. Here, low-activity customers (customers with less than 5 orders) will be excluded from the orders_products_all to generate a final dataframe. 

Firstly, a new variable should be created where the maximum number of orders = ("max_order_number") that each user ordered is calculated. This step would be described in 3.1 Aggregate variable. Secondly, a derived variable will be created to flag user as low and no-low under the "activitiy_customer" variable. This step would be described in 3.2 Derived variable. Finally in 3.3 Final dataframe, the "activity_customer" variable would be used to exclude observations that are not of the CFO's interest, generating out final dataframe. 

# 2. Importing data

## 2.1 Importing libraries

#Import analytical libraries
import pandas as pd
import numpy as np
import os

## 2.2 Importing data

#Project folder path into a string to easily retrieve data
path = r'C:\Users\Ich\Documents\12-2023 Instacart Basket Analysis'

### Order products all

#Import “orders_products_all_step4.pkl”
#See "Step 4 IC Orders products all" to check for merging details
orders_products_all = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared data', 'orders_products_all_step4.pkl'))

#Check df size
orders_products_all.shape

#Check headers
orders_products_all.head()

# 3. Exclusion flag

## 3.1 Aggregate variable

#Calculating the maximum number of orders = ("max_order_number") that each user ordered
orders_products_all['max_order_number'] = orders_products_all.groupby(['user_id'])['order_number'].transform(np.max)

#See results
orders_products_all.head()

#Check aggregations
#Snap view of maximum order number per user
orders_products_all.groupby(['user_id'])['order_number'].max()

#Creating a temporary df
df = orders_products_all[['user_id', 'max_order_number']] 

#Subgroupping user_id to check flag
df.loc[df['user_id'].isin([5, 206208])]
#Aggregation was calculated correctly

## 3.2 Derived variable

#Create flag
#Flag condition
#1. low-activity customers (customers with less than 5 orders)
#2. no-low activity customers (customers with 5 or more orders)
orders_products_all.loc[orders_products_all['max_order_number'] < 5, 'activity_customer'] = 'low'

orders_products_all.loc[orders_products_all['max_order_number'] >= 5, 'activity_customer'] = 'no_low'

#Check flags
#Snap view of flags
orders_products_all.groupby(['user_id'])['activity_customer'].max()

#Creating a temporary df
df = orders_products_all[['user_id', 'max_order_number', 'activity_customer']] 

#Subgroupping user_id to check flag
df.loc[df['user_id'].isin([5, 206208])]
#Flag is set correctly

#Print the frequency of flag "activity_customer"
orders_products_all['activity_customer'].value_counts(dropna = False) 
#It is expected that 30992966 observation will remain on the final df

## 3.3 Final dataframe

#Excluding observations using the flag
#Flag "activity_customer", exclude all observations labelled low-activity
orders_products_final = orders_products_all.loc[orders_products_all['activity_customer'] == 'no_low' ]  

#See results
orders_products_final.head()

#See results
#Print the frequency of flag "activity_customer"
orders_products_final['activity_customer'].value_counts(dropna = False) 
#Only no_low observations are included

#See results
orders_products_final.shape
#As it was expected, only 30992966 remained 

#Drop "activity_customer" variable
orders_products_final = orders_products_final.drop(columns = ['activity_customer'])

#check df size
orders_products_final.shape

#check df headers
orders_products_final.head()

# 4. Exporting data

#Exporting to prepared data folder
#The pickle format is preferred for large df. This df contains 31M rows
orders_products_final.to_pickle(os.path.join(path, '02 Data','Prepared data', 'orders_products_final_step5.pkl'))
